requirements.txt
flask
MySQL-python
pyyaml